# service/api/ban.unbanUser.go
package api

import (
	"net/http"
	"strconv"
	"strings"

	"github.com/MarinaGRey/WASA/service/api/reqcontext"
	"github.com/julienschmidt/httprouter"
)

func (rt *_router) unbanUser(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
	var banner User
	var banned User

	// check parameters
	userId, err := strconv.ParseUint(ps.ByName("UserId"), 10, 64)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	banUserId, err := strconv.ParseUint(ps.ByName("BanUserId"), 10, 64)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	// correct user
	requestUserId, err := strconv.ParseUint(strings.
		Split(r.Header.Get("Authorization"), " ")[1], 10, 64)
	if err != nil {
		http.Error(w, "Invalid authorization requestUserId", http.StatusBadRequest)
		return
	}

	if requestUserId != userId {
		w.WriteHeader(http.StatusUnauthorized)
		return
	}

	// cannot unban yourself
	if userId == banUserId {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	banner.UserId = userId
	banned.UserId = banUserId

	// unban is done
	err = rt.db.UnbanUser(banner.ToDatabase(), banned.ToDatabase())
	if err != nil {
		ctx.Logger.WithError(err).Error("unbanUser/db.UnbanUser: executing error")
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}